# Performance improvements for the trace SDK in `Span`.