Good day,
Thanks for review this PR!
This PR follows from https://github.com/open-telemetry/opentelemetry-go/pull/5864 and avoid multiple locks in `End`, `RecordError`, `AddEvent` and `AddLink`.

Benchstats result are:
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/trace
cpu: 11th Gen Intel(R) Core(TM) i5-11400H @ 2.70GHz
           │ event-3cbd9671.txt │        event-pr-eba0e961.txt        │
           │       sec/op       │   sec/op     vs base                │
SpanEnd-12          63.07n ± 1%   52.28n ± 1%  -17.11% (p=0.000 n=10)

           │ event-3cbd9671.txt │     event-pr-eba0e961.txt      │
           │        B/op        │    B/op     vs base            │
SpanEnd-12           0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

           │ event-3cbd9671.txt │     event-pr-eba0e961.txt      │
           │     allocs/op      │ allocs/op   vs base            │
SpanEnd-12           0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```